aura-skin-ai/backend/ai-engine/requirements.txt
fastapi
uvicorn
python-dotenv
opencv-python
mediapipe
numpy
scikit-image
pydantic
python-multipart
redis
rq
pillow
scikit-learn
httpx
supabase
